+++ src/main/kotlin/com/himanshoe/feature/frontend/FrontendRouting.kt
@@ -1,44 +1,17 @@
 package com.himanshoe.feature.frontend
 
+import com.himanshoe.base.provider.exception.ExceptionProvider
+import com.himanshoe.di.DomainLocator
+import com.himanshoe.feature.frontend.views.landingPage
+import com.himanshoe.feature.frontend.views.responsePage
 import io.ktor.application.*
-import io.ktor.html.*
 import io.ktor.routing.*
-import kotlinx.html.*
 
-fun Application.frontendRouting() {
+fun Application.frontendRouting(domainLocator: DomainLocator, provideExceptionProvider: ExceptionProvider) {
+
     routing {
-        get("/") {
-            val name = "Ktor"
-            call.respondHtml {
-                head {
-                    title {
-                        +"Welcome to URL Shortener"
-                    }
-                }
-                body {
-                    form(
-                        action = "/v1/form",
-                        method = FormMethod.post,
-                    ) {
-                        label {
-                            input(
-                                type = InputType.url,
-                                name = "url",
-                            ) {
-                                placeholder = "Enter your url"
-                                required = true
-                            }
-                        }
-
-                        button(
-                            classes = "secondary",
-                            type = ButtonType.submit,
-                        ) {
-                            +"Shorten"
-                        }
-                    }
-                }
-            }
-        }
+        landingPage()
+        responsePage(domainLocator,provideExceptionProvider)
     }
+
 }
